requirements.txt
streamlit
torch
transformers
sentence-transformers
spacy
pandas
numpy
scikit-learn
shap
requests
beautifulsoup4
pymongo
bcrypt
python-dotenv
plotly
email-validator
tiktoken
sentencepiece
protobuf
textstat
